+++ tests/testthat/test-grep_case.R
@@ -26,7 +26,7 @@ test_that("grep_case() .exhaustive", {
 
   expect_error(
     grep_case(words, "cat" ~ "feline", .exhaustive = TRUE),
-    "The following values were not matched: dogwood and dogma.",
+    'The following values were not matched: "dogwood" and "dogma".',
   )
 
   expect_equal(