+++ NEWS.md
@@ -1,5 +1,7 @@
 # incase (development version)
 
+* Improve formatting of error messages when `.exhaustive` is `TRUE`.
+
 # incase 0.4.0
 
 * The arguments `preserve`, `default`, and `ordered` have been deprecated in favor of `.preserve`, `.default`, and `.ordered` (#9).